make BasePrimitive2D easier extensible with VisitorParameters
BasePrimitive2D virtual methods get2DDecomposition and getB2DRange
are used in many subclasses of BasePrimitive2D. If we want to
extend the parameters that we want to pass to those classes (and
we will need to in the future changes) we need to change all the
subclasses too. So this commit intoduces VisitorParameters class,
which holds the parameters that we want to pass, which avoids the
need to extend the BasePrimitive2D API.

The only member of VisitorParameters is ViewInformation2D, which
was previously a parameter for both methods.

Change-Id: I39afc28707f1511aafce4e8a84dbc45b84fc8cd5