vcl: bring back RGB565 scanline transformer
While we don't support this as a Bitmap format anymore, we still
need to transform a buffer that is in RGB565 format in some cases.
For example backwards compatibility or if a certain bitmap format
supports such pixel format.
This change also simplifies some scanline transformers by removing
code duplication.

Change-Id: I64aa258b8b1fbebf0ed174c0d5fdd2f75f382b28