Fail module-deps.pl on more errors
* Fail on the error code returned from make
* Collect dependency errors in the tree and print them

The "perldoc -f open" has an example in the "Opening a filehandle
into a command" section, which uses waitpid in the parent. Since
we parse the pipe
I'm not sure the waitpid actually works, as the process has
already ended at this point and the STDOUT pipe from make is
closed. So waitpid actually returns -1 here, as the process
is already gone. From "perldoc -f open" it is not clear, what
exactly happens: if the child calls execve, then the exit
wouldn't make any sense. The STDOUT pipe is just
open in the parent process. The whole setup

Change-Id: I2e77aa33a0eaa5d6d3e06904bb0af53f4b66ef91